requirements.txt
django
djangorestframework
Pillow
django-base64field
